requirements.txt
ccxt>=4.3.0
ccxt[pro]>=4.3.0
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.26.0
xgboost>=2.0.0
scikit-learn>=1.4.0
lightgbm>=4.3.0
loguru>=0.7.0
rich>=13.7.0
pyyaml>=6.0.1
python-dotenv>=1.0.0
aiohttp>=3.9.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
joblib>=1.3.0
